requirements.txt
Django==1.9.8
PyJWT==1.4.2
cffi==1.5.2
cryptography==1.2.3
defusedxml==0.4.1
dj-database-url==0.4.0
django-allauth==0.24.1
django-fernet-fields==0.4
django-localflavor==1.3
django-nested-admin==3.0.10
django-rest-auth==0.7.0
djangorestframework==3.3.2
gunicorn==19.4.5
idna==2.1
oauthlib==1.1.2
psycopg2==2.6.1
pyasn1==0.1.9
pycparser==2.14
python-social-auth==0.2.14
python3-openid==3.0.10
requests==2.9.1
requests-oauthlib==0.6.2
six==1.10.0
whitenoise==2.0.6
